test: extract `schema::test_option` from `schema::test_generic_enums`
